pyproject.toml
[tool.poetry]
name = "easyremind"
version = "0.1.0"
description = ""
authors = ["daima3629 <daima3629@usbc.be>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
